requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
flake8==3.8.4
identify==1.5.13
idna==2.10
ipstack==0.1.4
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.1
pre-commit==2.10.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
